refactor: Decompose UI components for better separation of concern
- Improve code maintainability and reusability
- Reduce component complexity
